# src/main/java/com/org/model/ComparisionRequest.java
package com.org.model;

import javax.validation.constraints.NotBlank;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

@Builder
@Setter
@Getter
@ToString
@AllArgsConstructor
@NoArgsConstructor
@EqualsAndHashCode
@JsonInclude(JsonInclude.Include.NON_NULL)
public class ComparisionRequest {

  @NotBlank(message = "name1 cannot be empty or null")
  @JsonProperty("name1")
  private String name1;

  @NotBlank(message = "name2 cannot be empty or null")
  @JsonProperty("name2")
  private String name2;
}